rename tests merge -> match
